Update testing files
