--- src/main/java/Message.java
@@ -1,14 +1,0 @@
-public class Message {
-
-    private static final String greeting = "Hello! I'm Zenith\nWhat can I do for you?\n";
-    private static final String exit = "Bye. Hope to see you again soon!";
-
-    public static String getGreeting() {
-        return Message.greeting;
-    }
-
-    public static String getExit() {
-        return Message.exit;
-    }
-
-}
